Moved inspect functions into own module
